requirements.txt
beautifulsoup4
lxml
requests
psycopg2-binary
python-dotenv
clickhouse-connect    
dotenv
pandas
